Add files via upload
Question 1 Display the data types of each column using the function dtypes, then take a screenshot and submit it, include your code in the image.
Question 2 Drop the columns "id" and "Unnamed: 0" from axis 1 using the method drop(), then use the method describe() to obtain a statistical summary of the data. Take a screenshot and submit it, make sure the inplace parameter is set to True
Question 3 Use the method value_counts to count the number of houses with unique floor values, use the method .to_frame() to convert it to a dataframe.
Question 4 Use the function boxplot in the seaborn library to determine whether houses with a waterfront view or without a waterfront view have more price outliers.
Question 5 Use the function regplot in the seaborn library to determine if the feature sqft_above is negatively or positively correlated with price.
Question 6 Fit a linear regression model to predict the 'price' using the feature 'sqft_living' then calculate the R^2. Take a screenshot of your code and the value of the R^2
Question 7 Fit a linear regression model to predict the 'price' using the list of features:
Question 8 Use the list to create a pipeline object to predict the 'price', fit the object using the features in the list features, and calculate the R^2.
Question 9 Create and fit a Ridge regression object using the training data, set the regularization parameter to 0.1, and calculate the R^2 using the test data.
Question 10 Perform a second order polynomial transform on both the training data and testing data. Create and fit a Ridge regression object using the training data, set the regularisation parameter to 0.1, and calculate the R^2 utilising the test data provided. Take a screenshot of your code and the R^2.